+++ src/main.c
@@ -136,8 +136,8 @@ void process_metrics(GraphInterface* multigraph_1, GraphInterface* multigraph_2)
     printf("First graph with '%d' vertices and '%d' edges\n", multigraph_1->vertices, multigraph_1->calculate_size(multigraph_1));
     printf("Second graph with '%d' vertices and '%d' edges\n", multigraph_2->vertices, multigraph_2->calculate_size(multigraph_2));
 
-    double metric = default_algorithm.calculate_metric(multigraph_1, multigraph_1->vertices, multigraph_2, multigraph_2->vertices);
-    printf("Graph similarity metric between graphs: %.3f\n",metric);
+    int metric = default_algorithm.calculate_metric(multigraph_1, multigraph_1->vertices, multigraph_2, multigraph_2->vertices);
+    printf("Graph similarity metric between graphs: %d\n",metric);
 
     printf("------------------------------------------------\n\n");
 }
